-Updated SlimDX to latest released version ( June 2010 )

+++ Projects/Axiom/RenderSystems/DirectX9/D3DTextureManager.cs
@@ -139,7 +139,7 @@ namespace Axiom.RenderSystems.DirectX9
 		{
 			get
 			{
-				return device.AvailableTextureMemory;
+				return (int)device.AvailableTextureMemory;
 			}
 		}
 	}
